backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
groq
python-multipart
python-dotenv
sqlalchemy